requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-bigquery==3.38.0
google-cloud-tasks==2.19.3
google-cloud-pubsub==2.18.4
google-cloud-secret-manager==2.16.4
openai==2.4.0
httpx==0.28.1
aiohttp==3.9.1
pydantic==2.12.2
pydantic-settings==2.11.0
python-dotenv==1.0.0
asyncio-throttle==1.0.2
beautifulsoup4==4.12.3
lxml==5.1.0
google-generativeai==0.3.2
urllib3==2.0.7
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
